Add school admin panel
Loading the students poses a SQL injection threat since the school’s ID
is inserted into the SQL query without using prepared statements. This
is being left in the site to meet internal deadlines. It is justified
since exploiting the vulnerability requires a school ID. Teachers are
unlikely to attack the site and mischievous students that notice this
vulnerability are likely to be deterred by their lack of a school ID.